@@ -1,6 +1,6 @@
 module github.com/GoogleCloudPlatform/appengine/go11x/cloudsql
 
-go 1.22.0
+go 1.23.0
 
 require github.com/go-sql-driver/mysql v1.8.1
 
